refactor: view 컴포넌트로 분리 및 파티장/파티원별 보여지는 view 차이 구현
